Fix rgeo build commands
Previously, the rgeo build silently failed due to not being able to
link against required libraries. This is due to silently accepting but
ignoring the previously valid `--with-geos-dir` and friends.

At some point the option changed to `--with-opt-dir`, `—with-opt-lib`
and `--with-opt-include` plus needing `--with-geos-config`, and this
changes the build to reflect that.